Completion norms for 403 French sentences

Completion norms for 403 French sentences

brunel2025

Format

A data frame with 403 rows and 7 variables

target_word

The word (in French)

sentence

The sentence (in French), which is completed by target_word

target_word_emotionality

Categorical: Is the target word classified as "positive", "negative" or "neutral"

sentence_emotionality

Categorical: Is the sentence classified as "mixt", "negative", "neutral", "non specified", or "positive"

percentage_target_word

Numeric between 0 and 1; percentage of a test sample that identified the target word

valence_target_word

Numeric, average valence rating for the target word

arousal_target_word

Numeric, average arousal rating for the target word

Source

This data set was distributed as supplementary online materials to the paper by Brunel et al. (2025), and is available from the Open Science Repository via https://osf.io/7pc46. Retrieved on October 2, 2025. The data was published under a CC-By Attribution 4.0 International license, which permits redistribution. The anticlust packages includes the raw and unchanged data file as provided by Jeremy Brunel, and a script that converts the raw data into the package data.

References

Brunel, J., Dujardin, E., Delord, S. (2025) Emotional valence, cloze probability, and entropy: Completion norms for 403 French sentences. Behavior Research Methods 57(81). https://doi.org/10.3758/s13428-025-02604-7

Examples

groups <- anticlustering(
  brunel2025[, 3:7],
  K = 4, 
  objective = "kplus", 
  standardize = TRUE,
  method = "3phase"
)
# check out descriptive statistics for each variable for each of 4 groups:
# Numeric variables
mean_sd_tab(brunel2025[, 5:7], groups)
#>   percentage_target_word valence_target_word arousal_target_word
#> 1 "0.43 (0.32)"          "-0.27 (1.33)"      "3.05 (0.99)"      
#> 2 "0.43 (0.32)"          "-0.27 (1.33)"      "3.05 (0.99)"      
#> 3 "0.43 (0.32)"          "-0.28 (1.32)"      "3.05 (0.99)"      
#> 4 "0.43 (0.32)"          "-0.28 (1.33)"      "3.05 (0.99)"      
# Categorical variables:
table(groups, brunel2025$target_word_emotionality)
#>       
#> groups Negative Neutral Positive
#>      1       31      55       15
#>      2       31      55       15
#>      3       31      55       14
#>      4       31      56       14
table(groups, brunel2025$sentence_emotionality)
#>       
#> groups Mixt Negative Neutral Non Specified Positive
#>      1    1       16      70             3       11
#>      2    1       16      70             4       10
#>      3    1       15      71             3       10
#>      4    1       16      70             4       10