Objective value for the k-plus criterion

Objective value for the k-plus criterion

kplus_objective(x, clusters)

Arguments

x
A vector, matrix or data.frame of data points. Rows correspond to elements and columns correspond to features. A vector represents a single feature.

clusters
A vector representing (anti)clusters (e.g., returned by anticlustering or balanced_clustering)

Value

The value of the k-plus criterion.

Details

The k-plus criterion is an extension of the k-means criterion (i.e., the "variance", see variance_objective). The standard k-means objective is high if the means of the input variables are similar between clusters, but there is no guarantee that the standard deviations will also be similar (in fact, maximizing the k-means objective tends to decrease similarity in standard deviations in comparison to a completely random assignment). However, to achieve overall between-group similarity, it is desirable that the spread of the data is also similar between groups---and not just the means. This is accomplished by maximizing the k-plus criterion that also incorporates the standard deviations of the input variables.

Equalizing means and standard deviations simultaneously is accomplished by internally appending new variables to the data input x, one new variable for each column in x. These new variables contain the squared difference of each data point to the mean of the respective column, and are then included---in addition to the original data---in standard k-means anticlustering. This way, the average squared deviation of the data points to the means becomes similar between groups, which is the variance. Hence, the k-plus criterion simultaneously represents similarity in means and variance (and thus, the standard deviation), and can be used to simultaneously equalize the mean and the spread of the data.

Note

K-plus anticlustering has newly been implemented in the package anticlust (available since version 0.5.2). The author is currently working on a paper detailing the objective's background, but has already made the methodology available as its results have been very convincing thus far (e.g., check out the examples below). When using k-plus anticlustering in your research, it would be courteous to cite Papenberg and Klau (2020) as the primary anticlust reference, even though the criterion has not been described in that paper. In doubt, contact the author to inquire whether a new reference is available, or check out the package website (https://github.com/m-Py/anticlust).

References

Papenberg, M., & Klau, G. W. (2020). Using anticlustering to partition data sets into equivalent parts. Psychological Methods. Advance Online Publication. https://doi.org/10.1037/met0000301.

Author

Martin Papenberg martin.papenberg@hhu.de

Examples

data(schaper2019)
features <- schaper2019[, 3:6]

# Optimize k-plus criterion
kplus_groups <- anticlustering(
  features,
  K = 3,
  objective = "kplus"
)
# Optimize normal k-means criterion
kmeans_groups <- anticlustering(
  features,
  K = 3,
  objective = "variance"
)

# Compute k-plus criterion (k-plus is much better here)
kplus_objective(features, kplus_groups)

#> [1] 5433.657

kplus_objective(features, kmeans_groups)

#> [1] 5313.694

# Compare to k-means criterion (k-plus not much worse here)
variance_objective(features, kplus_groups)

#> [1] 628.1556

variance_objective(features, kmeans_groups)

#> [1] 628.1559

# Compare means and standard deviations after k-means and k-plus
# anticlustering (the standard deviations are usually much closer
# after k-plus anticlustering, but there is only little to no
# difference with regard to the means)
mean_sd_tab(features, kplus_groups)

#>   rating_consistent rating_inconsistent syllables     frequency     
#> 1 "4.49 (0.25)"     "1.10 (0.07)"       "3.44 (0.91)" "18.31 (2.42)"
#> 2 "4.49 (0.24)"     "1.10 (0.07)"       "3.41 (0.95)" "18.31 (2.40)"
#> 3 "4.49 (0.25)"     "1.10 (0.06)"       "3.41 (0.95)" "18.31 (2.42)"

mean_sd_tab(features, kmeans_groups)

#>   rating_consistent rating_inconsistent syllables     frequency     
#> 1 "4.49 (0.25)"     "1.10 (0.06)"       "3.41 (0.87)" "18.31 (2.09)"
#> 2 "4.49 (0.23)"     "1.10 (0.07)"       "3.41 (1.07)" "18.31 (2.67)"
#> 3 "4.49 (0.27)"     "1.10 (0.07)"       "3.44 (0.84)" "18.31 (2.44)"