An objective function measuring similarity of sets

Compute the discrepancy in means and standard deviations between clusters.

mean_sd_obj(features, clusters)

Arguments

features
A matrix or data.frame of data points. Rows correspond to elements and columns correspond to features.

clusters
A clustering vector

Value

A value quantifying similarity in means and standard deviations. Higher values indicate that means and standard deviations are more similar.

Details

This function can be passed as the argument objective to the function anticlustering to minimize differences in means and standard deviations between anticlusters.

Examples

data(schaper2019)
features <- schaper2019[1:48, 3:6]
ac <- anticlustering(
  features,
  K = 3,
  categories = schaper2019$room[1:48],
  objective = mean_sd_obj
)
by(features, ac, function(x) round(colMeans(x), 2))

#> ac: 1
#>   rating_consistent rating_inconsistent           syllables           frequency 
#>                4.48                1.12                3.38               18.06 
#> ------------------------------------------------------------ 
#> ac: 2
#>   rating_consistent rating_inconsistent           syllables           frequency 
#>                4.49                1.09                3.38               18.12 
#> ------------------------------------------------------------ 
#> ac: 3
#>   rating_consistent rating_inconsistent           syllables           frequency 
#>                4.50                1.11                3.31               18.19 

by(features, ac, function(x) round(apply(x, 2, sd), 2))

#> ac: 1
#>   rating_consistent rating_inconsistent           syllables           frequency 
#>                0.28                0.06                0.81                2.43 
#> ------------------------------------------------------------ 
#> ac: 2
#>   rating_consistent rating_inconsistent           syllables           frequency 
#>                0.21                0.07                0.89                2.42 
#> ------------------------------------------------------------ 
#> ac: 3
#>   rating_consistent rating_inconsistent           syllables           frequency 
#>                0.26                0.06                0.87                2.43