Maximize dispersion for K groups

Maximize dispersion for K groups

optimal_dispersion(
  x,
  K,
  solver = NULL,
  max_dispersion_considered = NULL,
  min_dispersion_considered = NULL,
  npartitions = 1,
  time_limit = NULL
)

Arguments

x

The data input. Can be one of two structures: (1) A feature matrix where rows correspond to elements and columns correspond to variables (a single numeric variable can be passed as a vector). (2) An N x N matrix dissimilarity matrix; can be an object of class dist (e.g., returned by dist or as.dist) or a matrix where the entries of the upper and lower triangular matrix represent pairwise dissimilarities.

K

The number of groups or a vector describing the size of each group.

solver

Optional argument; currently supports "lpSolve", "glpk", "symphony", and "gurobi". See optimal_anticlustering.

max_dispersion_considered

Optional argument used for early stopping. If the dispersion found is equal to or exceeds this value, a solution having the previous best dispersion is returned.

min_dispersion_considered

Optional argument used for speeding up the algorithm computation. If passed, the dispersion is optimized starting from this value instead the global minimum distance.

npartitions

The number of groupings that are returned, each having an optimal dispersion value (defaults to 1).

time_limit

Time limit in seconds, given to the solver. Default is there is no time limit.

Value

A list with four elements:

dispersion

The optimal dispersion

groups

An assignment of elements to groups (vector)

edges

A matrix of 2 columns. Each row contains the indices of elements that had to be investigated to find the dispersion (i.e., each pair of elements cannot be part of the same group in order to achieve maximum dispersion).

dispersions_considered

All distances that were tested until the dispersion was found.

Details

The dispersion is the minimum distance between two elements within the same group. This function implements an optimal method to maximize the dispersion. If the data input x is a feature matrix and not a dissimilarity matrix, the pairwise Euclidean distance is used. It uses the algorithm OptDispF presented Papenberg, Breuer, et al. (2025). It was originally developed as part of Max Diekhoff's Bachelor thesis at the Computer Science Department at Heinrich Heine University Düsseldorf.

To find out which items are not allowed to be grouped in the same cluster for maximum dispersion, the algorithm sequentially builds instances of a graph coloring problem, using an integer linear programming (ILP) representation. It is possible to specify the ILP solver via the argument solver (See optimal_anticlustering for more information on this argument). Optimally solving the maximum dispersion problem is NP-hard for K > 2 and therefore computationally infeasible for larger data sets. For K = 3 and K = 4, it seems that this approach scales up to several 100 elements, or even > 1000 for K = 3 (at least when using the Symphony or gurobi solver). For larger data sets, use the heuristic approaches in anticlustering or bicriterion_anticlustering. However, note that for K = 2, the optimal approach is usually much faster than the heuristics.

In the output, the element edges defines which elements must be in separate clusters in order to achieve maximum dispersion. All elements not listed here can be changed arbitrarily between clusters without reducing the dispersion. If the maximum possible dispersion corresponds to the minimum dispersion in the data set, the output elements edges and groups are set to NULL because all possible groupings have the same value of dispersion. In this case the output element dispersions_considered has length 1.

If a time_limit is set and the function cannot find in the optimal dispersion in the given time, it will throw an error.

Note

If the SYMPHONY solver is used, an unfortunate "message" is printed to the console when this function terminates:

sym_get_col_solution(): No solution has been stored!

This message is no reason to worry and instead is a direct result of the algorithm finding the optimal value for the dispersion. Unfortunately, this message is generated in the C code underlying the SYMPHONY library (via the printing function printf), which cannot be prevented in R.

References

Diekhoff (2023). Maximizing dispersion for anticlustering. Retrieved from https://www.cs.hhu.de/fileadmin/redaktion/Fakultaeten/Mathematisch-Naturwissenschaftliche_Fakultaet/Informatik/Algorithmische_Bioinformatik/Bachelor-_Masterarbeiten/2831963_ba_ifo_AbschlArbeit_klau_mapap102_madie120_20230203_1815.pdf

Papenberg, M., Breuer, M., Diekhoff, M., Tran, N. K., & Klau, G. W. (2025). Extending the Bicriterion Approach for Anticlustering: Exact and Hybrid Approaches. Psychometrika, 90(5), 1789–1808. https://doi.org/10.1017/psy.2025.10052

See also

Author

Max Diekhoff

Martin Papenberg martin.papenberg@hhu.de

Examples

N <- 30
M <- 5
K <- 3
data <- matrix(rnorm(N*M), ncol = M)
distances <- dist(data)

opt <- optimal_dispersion(distances, K = K)
opt
#> $dispersion
#> [1] 1.595477
#> 
#> $groups
#>  [1] 3 2 2 1 1 3 2 3 3 1 1 1 3 2 3 1 1 3 2 3 2 2 3 1 3 2 1 1 2 2
#> 
#> $groups_fixated
#>  [1] NA NA  2 NA  1 NA  2  3 NA NA  1  1 NA  2  3  1  1  3 NA  3  2  2  3  1  3
#> [26]  2  1 NA NA NA
#> 
#> $edges
#>       [,1] [,2]
#>  [1,]    5   23
#>  [2,]   20   22
#>  [3,]   21   25
#>  [4,]   17   22
#>  [5,]   11   20
#>  [6,]   11   23
#>  [7,]   23   26
#>  [8,]    5   26
#>  [9,]   14   24
#> [10,]   14   15
#> [11,]   17   20
#> [12,]    5   25
#> [13,]   11   22
#> [14,]    8   14
#> [15,]   12   23
#> [16,]    7   16
#> [17,]    3    8
#> [18,]   11   26
#> [19,]    7   27
#> [20,]    5   21
#> [21,]    3   18
#> [22,]   15   24
#> [23,]   12   26
#> [24,]    8   11
#> 
#> $dispersions_considered
#>  [1] 0.8928183 1.0434564 1.0531281 1.1222064 1.1345623 1.1380846 1.1525358
#>  [8] 1.1621927 1.1713164 1.1864504 1.1891757 1.2029536 1.3096835 1.3142788
#> [15] 1.4073222 1.4157117 1.4230152 1.4685792 1.4813877 1.5299284 1.5327486
#> [22] 1.5351956 1.5889212 1.5939599 1.5954767
#> 

# Compare to bicriterion heuristic:
groups_heuristic <- anticlustering(
  distances, 
  K = K,
  method = "brusco", 
  objective = "dispersion", 
  repetitions = 100
)
c(
  OPT = dispersion_objective(distances, opt$groups),
  HEURISTIC = dispersion_objective(distances, groups_heuristic)
)
#>       OPT HEURISTIC 
#>  1.595477  1.595477 

# Different group sizes are possible:
table(optimal_dispersion(distances, K = c(15, 10, 5))$groups)
#> 
#>  1  2  3 
#> 15 10  5 

# Induce cannot-link constraints by maximizing the dispersion:
solvable <- matrix(1, ncol = 6, nrow = 6)
solvable[2, 1] <- -1
solvable[3, 1] <- -1
solvable[4, 1] <- -1
solvable <- as.dist(solvable)
solvable
#>    1  2  3  4  5
#> 2 -1            
#> 3 -1  1         
#> 4 -1  1  1      
#> 5  1  1  1  1   
#> 6  1  1  1  1  1

# An optimal solution has to put item 1 in a different group than 
# items 2, 3 and 4 -> this is possible for K = 2
optimal_dispersion(solvable, K = 2)$groups
#> [1] 2 1 1 1 2 2

# It no longer works when item 1 can also not be linked with item 5:
# (check out output!)
unsolvable <- as.matrix(solvable)
unsolvable[5, 1] <- -1
unsolvable <- as.dist(unsolvable)
unsolvable
#>    1  2  3  4  5
#> 2 -1            
#> 3 -1  1         
#> 4 -1  1  1      
#> 5 -1  1  1  1   
#> 6  1  1  1  1  1
optimal_dispersion(unsolvable, K = 2)
#> $dispersion
#> [1] -1
#> 
#> $groups
#> NULL
#> 
#> $edges
#> NULL
#> 
#> $dispersions_considered
#> [1] -1
#> 
# But:
optimal_dispersion(unsolvable, K = c(2, 4)) # group sizes, not number of groups
#> $dispersion
#> [1] 1
#> 
#> $groups
#> [1] 2 1 1 1 1 2
#> 
#> $groups_fixated
#> [1]  2  1  1  1  1 NA
#> 
#> $edges
#>      [,1] [,2]
#> [1,]    1    2
#> [2,]    1    3
#> [3,]    1    4
#> [4,]    1    5
#> 
#> $dispersions_considered
#> [1] -1  1
#>